requirements.txt
ax-platform==0.1.10
black==19.10b0
opencv-python==4.2.0.32
kaggle==1.5.6
matplotlib==3.1.3
numpy==1.17.4
pandas==0.25.1
pillow==6.2.0
scikit-learn==0.21.3
torch==1.4.0
torchvision==0.5.0
